requirements.txt
pyrogram
tgcrypto
ytthumb
python-dotenv
